Moved qa functions to new file
